搞完Spring声明式事务

+++ boot-datasource/boot-datasource-jpa/src/main/java/com/msr/better/jpa/service/impl/StudentServiceImpl.java
@@ -4,7 +4,10 @@ import com.msr.better.jpa.dao.StudentRepository;
 import com.msr.better.jpa.domain.Student;
 import com.msr.better.jpa.service.IStudentService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.ApplicationContext;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Isolation;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -20,6 +23,10 @@ public class StudentServiceImpl implements IStudentService {
     @Autowired
     private StudentRepository studentRepository;
 
+    @Autowired
+    private ApplicationContext context;
+
+    @Transactional(rollbackFor = RuntimeException.class, propagation = Propagation.REQUIRES_NEW)
     @Override
     public Student findStudentById(Long id) {
         return studentRepository.getOne(id);
@@ -30,7 +37,7 @@ public class StudentServiceImpl implements IStudentService {
         return studentRepository.findByNameLike(name);
     }
 
-    @Transactional(rollbackFor = RuntimeException.class)
+    @Transactional(rollbackFor = RuntimeException.class, isolation = Isolation.REPEATABLE_READ, propagation = Propagation.REQUIRES_NEW)
     @Override
     public Student insertStudent(Student student) {
         return studentRepository.save(student);
@@ -42,9 +49,14 @@ public class StudentServiceImpl implements IStudentService {
         return studentRepository.save(student);
     }
 
-    @Transactional(rollbackFor = RuntimeException.class)
+    @Transactional(rollbackFor = RuntimeException.class, isolation = Isolation.REPEATABLE_READ)
     @Override
     public void deleteStudent(Long id) {
+        IStudentService studentService = context.getBean(IStudentService.class);
+        Student db = findStudentById(id);
+        if (db == null) {
+            throw new RuntimeException("数据不存在");
+        }
         studentRepository.deleteById(id);
     }
